Fixed everything.
Drawing the circle now adds it to the circles array, so all the circles in the array have been drawn.

Exists properly returns if the circle exists in that location.

Rounding the values of the intersection function to make sure that repeated circles (or that the difference in position of x and y is > than 0 and < than 1